Move camera and microphone examples to sample module #28 camera and microphone screens moved to sample module
