+++ sample/src/main/res/drawable/camera.webp
