+++ sample/src/main/res/drawable/ic_warning.xml
