+++ sample/src/main/res/drawable/microphone.png
